layer0_web: Handle all non-word characters
In the rds app for layer0_web we replace both dash and dot with
underscore[0], we don't do that in the corresponding chart, resulting in a
bug where the environment variable created will not match the one that
is actually looked for in the application.

This patch will replace[1] all non word characters[2] with underscore (i.e.
only [a-zA-Z0-9_] will remain).

I will shortly submit a patch for the rds application as well, to make
sure we use the same scheme in both places to avoid further bugs like
this.

Notes:
 0. https://github.com/Sciebo-RDS/Sciebo-RDS/blob/develop/RDS/layer0_ingress/web/server/src/app.py#L107
 1. https://github.com/Masterminds/sprig/blob/master/docs/strings.md#regexreplaceall-mustregexreplaceall
 2. https://yourbasic.org/golang/regexp-cheat-sheet/#character-classes